requirements.txt
fastapi
uvicorn
sqlalchemy
mysqlclient
pydantic
python-dotenv
passlib
python-jose
starlette
cryptography
bcrypt
jose

